refactor digest-related intrinstics
